services/log-ingestor/requirements.txt
fastapi
uvicorn
kafka-python
pydantic
prometheus-client
